feat(view): implement method read bridge size
