+++ docs/README.md
@@ -1,10 +1,10 @@
 ## 기능목록📋
 
-1. [ ] 다리 길이를 입력받는다.
+1. [X] 다리 길이를 입력받는다.
 
 
 2. [ ] 다리 길이 입력값을 검증한다.
-   1. [ ] 숫자인지 검증한다.
+   1. [X] 숫자인지 검증한다.
    2. [ ] 3 이상 20 이하의 숫자인지 검증한다.
